+ Redid the test level geometry in Blender rather than Pro Buildier. + Added the new mage tower. When a monster wave starts, a villager will go to and occupy an empty mage tower, appearing on top of it as a mage. Fires magic homing projectiles at monsters. + Created the first projectiles in the game. + When a building is attacked, villagers will no longer respond to it's distress call if they are already in combat. + Fixed some bugs.

+++ Assets/Scripts/AI/Monsters/IMonster.cs
@@ -19,11 +19,11 @@ public interface IMonster
 
     bool IsAttacking { get; }
 
-    int GetDangerValue();
+    float GetDangerValue();
     int GetScoreValue();
 
 
-    int GetAttackPower();
+    float GetAttackPower();
     int GetTier();
     
 }
